requirements.txt
numpy>=1.19.4
future>=0.18.2
Pillow>=8.0.1